src/pyproject.toml
[tool.poetry]
name = "machine-learning-project"
version = "0.1.0"
description = ""
authors = ["Felix Böhm <felix@felixboehm.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.0"
jupyterlab = "^3.0.16"
google-cloud-translate = "2.0.1"
six = "^1.16.0"
spacy = "^3.0.6"
scikit-learn = "^0.24.2"
pyarrow = "^4.0.1"
matplotlib = "^3.4.2"
langdetect = "^1.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
